New Metrics Module
git-svn-id: https://bigstreams.googlecode.com/svn/trunk@21 e9027e87-718f-2365-ccfc-279084edcb2a

+++ commons/src/main/java/org/streams/commons/file/SyncPointer.java
@@ -7,17 +7,19 @@ package org.streams.commons.file;
  */
 public class SyncPointer implements Comparable<SyncPointer>{
 
-	long lockId;
+	int lockId;
 	long filePointer;
 	int linePointer;
 	long fileSize;
 	
-	public long getLockId() {
+	public int getLockId() {
 		return lockId;
 	}
-	public void setLockId(long lockId) {
+	
+	public void setLockId(int lockId) {
 		this.lockId = lockId;
 	}
+	
 	public long getFilePointer() {
 		return filePointer;
 	}
